move all manifest related code into one place
